requirements.txt
fastapi==0.115.12
Flask_AppBuilder==4.5.2
gspread==6.2.1
joblib==1.4.2
keras==3.9.2
mlflow==2.22.0
mlflow_skinny==2.22.0
numpy==2.1.3
oauth2client==4.1.3
pandas==2.2.3
plotly==6.0.1
psycopg2-binary==2.9.10
Requests==2.32.3
scikit_learn==1.6.1
SQLAlchemy==1.4.54
streamlit==1.45.0
tensorflow_cpu==2.19.0
boto3
